feat(load): check whether the page is loaded successfully
Check whether the page is loaded successfully (Keep the compatibility if the browser's `load` event has been triggered).
